src/jp/co/internous/dto/LoginDTOTest.java
package jp.co.internous.dto;

import static org.junit.Assert.*;

import org.junit.Test;

public class LoginDTOTest {

	@Test
	public void testGetuserID() {
		String userID ="aaa";
		LoginDTO expected = new LoginDTO();
		expected.setUserID(userID);
		assertSame(userID,LoginDTO.getUserID());
		assertEquals(userID,LoginDTO.getUserID());

	}
	@Test
	public void testSetUserID() {
		String userID ="aaa";
		LoginDTO expected = new LoginDTO();
		expected.setUserID("aaa");
		String value = LoginDTO.getUserID();
		assertSame(value,userID);
		assertEquals(value,userID);

	}

	@Test
	public void testGetPassword() {
		String password ="1234";
		LoginDTO expected = new LoginDTO();
		expected.setPassword(password);
		assertSame(password,LoginDTO.getPassword());
		assertEquals(password,LoginDTO.getPassword());

	}
	@Test
	public void testSetPassword() {
		String password ="1234";
		LoginDTO expected = new LoginDTO();
		expected.setPassword("1234");
		String value = LoginDTO.getPassword();
		assertSame(value,password);
		assertEquals(value,password);

	}
}
